initial rough refactor
