+++ test_mq_service.py
@@ -18,9 +18,9 @@ from app import db
 from app.exceptions import InventoryException
 from app.exceptions import ValidationException
 from app.models import Host
-from app.queue.ingress import _validate_json_object_for_utf8
-from app.queue.ingress import event_loop
-from app.queue.ingress import handle_message
+from app.queue.queue import _validate_json_object_for_utf8
+from app.queue.queue import event_loop
+from app.queue.queue import handle_message
 from lib.host_repository import AddHostResults
 from test_utils import rename_host_table_and_indexes
 from test_utils import valid_system_profile
@@ -32,7 +32,12 @@ class MockEventProducer:
         self.key = None
         self.headers = None
 
-    def write_event(self, event, key, headers):
+    def write_event_egress_topic(self, event, key, headers):
+        self.event = event
+        self.key = key
+        self.headers = headers
+
+    def write_event_events_topic(self, event, key, headers):
         self.event = event
         self.key = key
         self.headers = headers
@@ -85,7 +90,7 @@ class MQServiceTestCase(MQServiceBaseTestCase):
 
         mock_event_producer.assert_not_called()
 
-    @patch("app.queue.egress.datetime", **{"now.return_value": datetime.now(timezone.utc)})
+    @patch("app.queue.events.datetime", **{"now.return_value": datetime.now(timezone.utc)})
     def test_handle_message_happy_path(self, datetime_mock):
         expected_insights_id = str(uuid.uuid4())
         host_id = uuid.uuid4()
@@ -100,15 +105,15 @@ class MQServiceTestCase(MQServiceBaseTestCase):
             },
         }
         with self.app.app_context():
-            with unittest.mock.patch("app.queue.ingress.host_repository.add_host") as m:
+            with unittest.mock.patch("app.queue.queue.host_repository.add_host") as m:
                 m.return_value = ({"id": host_id, "insights_id": None}, AddHostResults.created)
                 mock_event_producer = Mock()
                 handle_message(json.dumps(message), mock_event_producer)
 
-                mock_event_producer.write_event.assert_called_once()
+                mock_event_producer.write_event_egress_topic.assert_called_once()
 
                 self.assertEquals(
-                    json.loads(mock_event_producer.write_event.call_args[0][0]),
+                    json.loads(mock_event_producer.write_event_egress_topic.call_args[0][0]),
                     {
                         "host": {"id": str(host_id), "insights_id": None},
                         "platform_metadata": {},
@@ -142,8 +147,8 @@ class MQServiceTestCase(MQServiceBaseTestCase):
     #     pass
 
 
-@patch("app.queue.ingress.build_egress_topic_event")
-@patch("app.queue.ingress.add_host", return_value=(MagicMock(), None))
+@patch("app.queue.queue.build_egress_topic_event")
+@patch("app.queue.queue.add_host", return_value=(MagicMock(), None))
 class MQServiceParseMessageTestCase(MQServiceBaseTestCase):
     def _message(self, display_name):
         return f'{{"operation": "", "data": {{"display_name": "hello{display_name}"}}}}'
@@ -246,7 +251,7 @@ class MQhandleMessageTestCase(MQAddHostBaseClass):
         event = json.loads(mock_event_producer.event)
         self.assertEqual(event["platform_metadata"], metadata)
 
-    @patch("app.queue.ingress.host_repository.add_host")
+    @patch("app.queue.queue.host_repository.add_host")
     def test_handle_message_verify_message_key_and_metadata_not_required(self, add_host):
         host_data = self._host_data()
         add_host.return_value = (host_data, AddHostResults.created)
@@ -261,7 +266,7 @@ class MQhandleMessageTestCase(MQAddHostBaseClass):
         event = json.loads(mock_event_producer.event)
         self.assertEqual(event["host"], host_data)
 
-    @patch("app.queue.ingress.host_repository.add_host")
+    @patch("app.queue.queue.host_repository.add_host")
     def test_handle_message_verify_message_headers(self, add_host):
         host_data = self._host_data()
 
@@ -280,7 +285,7 @@ class MQhandleMessageTestCase(MQAddHostBaseClass):
 
 
 class MQAddHostTestCase(MQAddHostBaseClass):
-    @patch("app.queue.egress.datetime", **{"now.return_value": datetime.now(timezone.utc)})
+    @patch("app.queue.events.datetime", **{"now.return_value": datetime.now(timezone.utc)})
     def test_add_host_simple(self, datetime_mock):
         """
         Tests adding a host with some simple data
@@ -307,7 +312,7 @@ class MQAddHostTestCase(MQAddHostBaseClass):
 
         self._base_add_host_test(host_data, expected_results, host_keys_to_check)
 
-    @patch("app.queue.egress.datetime", **{"now.return_value": datetime.now(timezone.utc)})
+    @patch("app.queue.events.datetime", **{"now.return_value": datetime.now(timezone.utc)})
     def test_add_host_with_system_profile(self, datetime_mock):
         """
          Tests adding a host with message containing system profile
@@ -335,7 +340,7 @@ class MQAddHostTestCase(MQAddHostBaseClass):
 
         self._base_add_host_test(host_data, expected_results, host_keys_to_check)
 
-    @patch("app.queue.egress.datetime", **{"now.return_value": datetime.now(timezone.utc)})
+    @patch("app.queue.events.datetime", **{"now.return_value": datetime.now(timezone.utc)})
     def test_add_host_with_tags(self, datetime_mock):
         """
          Tests adding a host with message containing tags
@@ -396,7 +401,7 @@ class MQAddHostTestCase(MQAddHostBaseClass):
                 with self.assertRaises(ValidationException):
                     self._handle_message(host_data)
 
-    @patch("app.queue.egress.datetime", **{"utcnow.return_value": datetime.utcnow()})
+    @patch("app.queue.events.datetime", **{"utcnow.return_value": datetime.utcnow()})
     def test_add_host_empty_keys_system_profile(self, datetime_mock):
         insights_id = str(uuid.uuid4())
 
@@ -415,7 +420,7 @@ class MQAddHostTestCase(MQAddHostBaseClass):
             with self.assertRaises(ValidationException):
                 handle_message(json.dumps(message), mock_event_producer)
 
-    @patch("app.queue.egress.datetime", **{"utcnow.return_value": datetime.utcnow()})
+    @patch("app.queue.events.datetime", **{"utcnow.return_value": datetime.utcnow()})
     def test_add_host_empty_keys_facts(self, datetime_mock):
         insights_id = str(uuid.uuid4())
 
@@ -442,7 +447,7 @@ class MQAddHostTestCase(MQAddHostBaseClass):
                     with self.assertRaises(ValidationException):
                         handle_message(json.dumps(message), mock_event_producer)
 
-    @patch("app.queue.egress.datetime", **{"utcnow.return_value": datetime.utcnow()})
+    @patch("app.queue.events.datetime", **{"utcnow.return_value": datetime.utcnow()})
     def test_add_host_with_invalid_stale_timestmap(self, datetime_mock):
         mock_event_producer = MockEventProducer()
 
@@ -525,7 +530,7 @@ class MQUpdateHostTestCase(MQAddHostBaseClass):
 
 
 class MQCullingTests(MQAddHostBaseClass):
-    @patch("app.queue.egress.datetime", **{"now.return_value": datetime.now(timezone.utc)})
+    @patch("app.queue.events.datetime", **{"now.return_value": datetime.now(timezone.utc)})
     def test_add_host_stale_timestamp(self, datetime_mock):
         """
         Tests to see if the host is succesfully created with both reporter
@@ -618,12 +623,12 @@ class MQValidateJsonObjectForUtf8TestCase(TestCase):
         with self.assertRaises(UnicodeEncodeError):
             _validate_json_object_for_utf8("hello\udce2\udce2")
 
-    @patch("app.queue.ingress._validate_json_object_for_utf8")
+    @patch("app.queue.queue._validate_json_object_for_utf8")
     def test_dicts_are_traversed(self, mock):
         _validate_json_object_for_utf8({"first": "item", "second": "value"})
         mock.assert_has_calls((call("first"), call("item"), call("second"), call("value")), any_order=True)
 
-    @patch("app.queue.ingress._validate_json_object_for_utf8")
+    @patch("app.queue.queue._validate_json_object_for_utf8")
     def test_lists_are_traversed(self, mock):
         _validate_json_object_for_utf8(["first", "second"])
         mock.assert_has_calls((call("first"), call("second")), any_order=True)
